requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
email-validator==2.2.0
WTForms==3.1.2
bcrypt==4.1.3
python-dotenv==1.0.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.8.4
seaborn==0.13.2
opencv-python==4.10.0.84
gunicorn==22.0.0
tensorflow-cpu==2.16.2
openai==1.51.0
google-generativeai==0.8.2
reportlab==4.0.9
weasyprint==62.3

